(Relationship) Between user and job.

+++ src/jobs/jobs.service.ts
@@ -1,10 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { JobCategory } from './models/job-category.enum';
 import { CreateJobDTO } from './dtos/create-job.dto';
 import { GetJobsFilterDTO } from './dtos/get-jobs-filter.dto';
 import { JobRepository } from './models/job.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Job } from './models/job.entity';
+import { User } from 'src/auth/models/user.entity';
 
 @Injectable()
 export class JobsService {
@@ -24,12 +29,15 @@ export class JobsService {
     return this.jobRepository.getJobs(filterDTO);
   }
 
-  async createJob(createJobDTO: CreateJobDTO): Promise<Job> {
-    return this.jobRepository.createJob(createJobDTO);
+  async createJob(createJobDTO: CreateJobDTO, user: User): Promise<Job> {
+    return this.jobRepository.createJob(createJobDTO, user);
   }
 
-  async deleteJob(id: number): Promise<void> {
-    const result = await this.jobRepository.delete(id);
+  async deleteJob(id: number, user: User): Promise<void> {
+    const job = await this.getJobById(id);
+    if (job.userId !== user.id)
+      throw new UnauthorizedException('You do not own this Job!');
+    const result = await this.jobRepository.delete({ id, userId: user.id });
     if (result.affected === 0)
       throw new NotFoundException(`Job with ID "${id}" not found!`);
   }
@@ -38,8 +46,11 @@ export class JobsService {
     id: number,
     title: string,
     category: JobCategory,
+    user: User,
   ): Promise<Job> {
     const job = await this.getJobById(id);
+    if (job.userId !== user.id)
+      throw new UnauthorizedException('You do not own this Job!');
     job.title = title;
     job.category = category;
     await job.save();
